+++ meta_package_manager/cli.py
@@ -21,6 +21,7 @@
 from __future__ import absolute_import, division, print_function
 
 import logging
+from json import dumps as json_dumps
 from operator import itemgetter
 
 import click
@@ -30,11 +31,18 @@ from tabulate import tabulate
 from . import __version__, logger
 from .managers import pool
 
-# Mapping of table formatiing options to tabulate's parameters.
-TABLE_FORMAT = {
+# Output rendering modes. From machine to human-readable.
+RENDERING_MODES = {
+    'json': 'json',
+    # Mapping of table formating options to tabulate's parameters.
     'plain': 'plain',
     'simple': 'simple',
     'fancy': 'fancy_grid'}
+
+
+def json(data):
+    """ Utility function to render data structure into pretty printed JSON. """
+    return json_dumps(data, sort_keys=True, indent=4, separators=(',', ': '))
 
 
 @click.group(invoke_without_command=True)
@@ -43,15 +51,15 @@ TABLE_FORMAT = {
     default='INFO', metavar='LEVEL',
     help='Either CRITICAL, ERROR, WARNING, INFO or DEBUG. Defaults to INFO.')
 @click.option(
-    '-t', '--table-format', type=click.Choice(TABLE_FORMAT), default='fancy',
-    help="Rendering format of tables. Defaults to fancy.")
+    '-o', '--output-format', type=click.Choice(RENDERING_MODES),
+    default='fancy', help="Rendering mode of the format. Defaults to fancy.")
 @click.version_option(__version__)
 @click.pass_context
-def cli(ctx, table_format):
+def cli(ctx, output_format):
     """ CLI for multi-package manager updates and upgrades. """
     level = click_log.get_level()
-    logger.debug('Verbosity set to {}.'.format(
-        logging._levelNames.get(level, level)))
+    level_name = logging._levelNames.get(level, level)
+    logger.debug('Verbosity set to {}.'.format(level_name))
 
     # Print help screen and exit if no sub-commands provided.
     if ctx.invoked_subcommand is None:
@@ -60,7 +68,11 @@ def cli(ctx, table_format):
 
     # Load up global options to the context.
     ctx.obj = {
-        'table_format': TABLE_FORMAT[table_format]}
+        'rendering': RENDERING_MODES[output_format]}
+
+    # Silent all log message in JSON rendering mode unless it's at debug level.
+    if ctx.obj['rendering'] == 'json' and level_name != 'DEBUG':
+        click_log.set_level(logging.CRITICAL * 2)
 
 
 @cli.command(short_help='List supported package managers and their location.')
@@ -68,9 +80,21 @@ def cli(ctx, table_format):
 def managers(ctx):
     """ List all supported package managers and their presence on the system.
     """
-    table = []
+    rendering = ctx.obj['rendering']
 
-    # Filters-out inactive managers.
+    # Machine-friendly data rendering.
+    if rendering == 'json':
+        fields = [
+            'name', 'id', 'cli_path', 'exists', 'executable', 'version_string',
+            'supported', 'available']
+        # JSON mode use print to output data because the logger is disabled.
+        print(json({
+            manager_id: {fid: getattr(manager, fid) for fid in fields}
+            for manager_id, manager in pool().items()}))
+        return
+
+    # Human-friendly content rendering.
+    table = []
     for manager_id, manager in pool().items():
         table.append([
             manager.name,
@@ -83,11 +107,9 @@ def managers(ctx):
                 manager.version if manager.version else '')
             if manager.exists else ''])
     table = [[
-        'Package manager', 'ID', 'CLI path', 'Found', 'Executable', 'Version'
-    ]] + sorted(table, key=itemgetter(1))
-
-    logger.info(tabulate(
-        table, tablefmt=ctx.obj['table_format'], headers='firstrow'))
+        'Package manager', 'ID', 'CLI path', 'Found', 'Executable',
+        'Version']] + sorted(table, key=itemgetter(1))
+    logger.info(tabulate(table, tablefmt=rendering, headers='firstrow'))
 
 
 @cli.command(short_help='Sync local package info.')
@@ -108,7 +130,8 @@ def sync(ctx):
 @click.pass_context
 def outdated(ctx):
     """ List available package updates and their versions for each manager. """
-    table = []
+    # Build-up a global list of outdated packages per manager.
+    outdated = {}
 
     for manager_id, manager in pool().items():
 
@@ -117,31 +140,50 @@ def outdated(ctx):
             logger.warning('Skip unavailable {} manager.'.format(manager_id))
             continue
 
+        # Force a sync to get the freshest updates.
         manager.sync()
 
         if manager.error:
             logger.error(manager.error)
 
-        for pkg_info in manager.updates:
-            table.append([
-                pkg_info['name'],
-                pkg_info['name'],
-                manager_id,
-                pkg_info['installed_version'],
-                pkg_info['latest_version'],
-            ])
+        outdated[manager_id] = {
+            'id': manager_id,
+            'name': manager.name,
+            'packages': [{
+                'name': pkg_info['name'],
+                'id': pkg_info['name'],
+                'installed_version': pkg_info['installed_version'],
+                'latest_version': pkg_info['latest_version']}
+                for pkg_info in manager.updates]}
+
+    rendering = ctx.obj['rendering']
+
+    # Machine-friendly data rendering.
+    if rendering == 'json':
+        # JSON mode use print to output data because the logger is disabled.
+        print(json(outdated))
+        return
+
+    # Human-friendly content rendering.
+    table = []
+    for manager_id, outdated_pkg in outdated.items():
+        table += [[
+            pkg_info['name'],
+            pkg_info['id'],
+            manager_id,
+            pkg_info['installed_version'],
+            pkg_info['latest_version']]
+            for pkg_info in outdated_pkg['packages']]
 
     # Sort table by package ID, then manager ID.
+    # TODO: ignore package ID case on sort.
     table = [[
         'Package name', 'ID', 'Manager', 'Installed version',
         'Latest version']] + sorted(table, key=itemgetter(1, 2))
-    logger.info(tabulate(
-        table, tablefmt=ctx.obj['table_format'], headers='firstrow'))
-
+    logger.info(tabulate(table, tablefmt=rendering, headers='firstrow'))
     # Print statistics.
     manager_stats = {
-        manager_id: len(manager.updates)
-        for manager_id, manager in pool().items() if manager.updates}
+        manager_id: len(packages) for manager_id, packages in outdated.items()}
     total_outdated = sum(manager_stats.values())
     per_manager_totals = ', '.join([
         '{} from {}'.format(v, k) for k, v in sorted(manager_stats.items())])