+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 `1.10.1 (unreleased) <https://github.com/kdeldycke/meta-package-manager/compare/v1.10.0...develop>`_
 ----------------------------------------------------------------------------------------------------
 
-* No changes yet.
+* Allow rendering of output data into ``json``.
 
 
 `1.10.0 (2016-10-04) <https://github.com/kdeldycke/meta-package-manager/compare/v1.9.0...develop>`_